Minor clean-up
- Moved environment to its own namespace
- Extracting Futures-related functions to a namespace

I'd like to be able to apply more macro-fu, but it doesn't look like I'll
get to. This depends too much on js code, and macros need to live in clj
files. figwheel throws a fit.

It doesn't look like I can use `apply` to clean some calls up for the same
reason.